requirements.txt
openai 
python-dotenv
pyinstaller